refact: created error handling lib; docs: created code documentation; clean adjusted spacing, grouping for readability
